module scroll-animated renamed to animate-scroll and transformed in a class | method 'animaScroll' renamed to 'anime' and binded to 'this'.
